store EPSG unit factors as f64
linear_unit_m and angular_unit_rad were f32, which cut factors such as the
US survey foot (0.30480060960121924) to ~7 digits and made the generator
write f64-precision text into f32 literals (2,190 clippy warnings). They are
now f64, written with Python's shortest round-trip repr, so they match EPSG
exactly. xeibe-geom's decimal_f64 workaround is gone. area_wgs84 stays f32.

verify_crs_tables.py parsed literals with rstrip("f32"), which strips any
trailing f/3/2 characters (53.32f32 -> 53.); it now removes the suffix only.

+++ crates/xeibe-crs/src/lib.rs
@@ -18,9 +18,8 @@
 use std::collections::HashMap;
 use std::sync::{Mutex, OnceLock};
 
-// Generated data: bbox values such as 3.14 are not π, and f32 literals keep
-// the digits EPSG publishes.
-#[allow(clippy::approx_constant, clippy::excessive_precision)]
+// Generated data: bbox longitudes such as 3.14 are not π.
+#[allow(clippy::approx_constant)]
 mod table;
 
 pub use table::{CRS, EPSG_DATE, EPSG_VERSION, PROJJSON_COUNT};
@@ -74,10 +73,10 @@ pub struct CrsRecord {
     pub dimension: u8,
     /// Size of the first axis's unit in metres, for CRSs measured in length
     /// (1.0 for metre, 0.3048… for foot). Used for `ArcByCenterPoint` radii.
-    pub linear_unit_m: Option<f32>,
+    pub linear_unit_m: Option<f64>,
     /// Size of the first axis's unit in radians, for CRSs measured in angle.
     /// Distinguishes a CRS in gradians from one in degrees.
-    pub angular_unit_rad: Option<f32>,
+    pub angular_unit_rad: Option<f64>,
     /// EPSG's published area of use, in WGS 84 degrees, as
     /// `[south, west, north, east]`.
     ///
